requirements.txt
langchain==1.2.10
langchain-core==1.2.11
langchain-openai==1.1.9
langchain-google-genai==4.2.1
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langgraph==1.0.8
python-dotenv==1.2.1
newrelic==11.5.0
